+++ lib/schema_registry_client/output/json_schema.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class SchemaRegistry
+module SchemaRegistry
   module Output
     module JsonSchema
       class << self