+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+# 0.0.7 - 2026-01-05
+
+* Switch to using `SchemaRegistry::Client` instead of bare `SchemaRegistry`.
+* Throw correct error when Avro schema is not found.
+
 # 0.0.6 - 2026-01-02
 
 * Initial release.